Startup a keep-alive timer to the ClientAccount too
