+++ univocity-trader-binance/src/main/java/com/univocity/trader/exchange/binance/BinanceClientAccount.java
@@ -49,6 +49,7 @@ class BinanceClientAccount implements ClientAccount {
 
 		factory = BinanceApiClientFactory.newInstance(apiKey, secret, asyncHttpClient);
 		client = factory.newRestClient();
+		new KeepAliveUserDataStream(client).start();
 	}
 
 	public double getMinimumBnbAmountToKeep() {
